Highlight winning boards
To achieve this, after each winning move on a minor board we check the winner of the main board and save it in the state. Then we can create a subscription that returns the winner of the main board, so that we can cheaply compare the game winner with the minor board winner.

Also, after the game is finished, all cells get inactive and the cell clicked handler essentially does nothing.